requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dnspython==1.16.0
idna==2.8
pushbullet.py==0.11.0
python-etcd==0.4.5
python-magic==0.4.15
python-twitch-client==0.6.0
requests==2.21.0
Rx==3.0.0a3
six==1.12.0
texttable==1.6.1
urllib3==1.24.1
websocket-client==0.56.0
